Move markExtend to commands
